Started rewrite of worship system.
Mostly for future compatibility. Interacting with an Idol of the chosen Deity will set that as your current deity.
TODO: Make a dynamic leveling system that checks for unlocked research.
TODO: Make quests (and slug eating) not increase your level directly, but instead check for required research and grant you new research when you interact with the Idol, this research being required to gain the powers.